Validate contract id route params
